기본 동작 완료.

+++ app/src/main/java/com/leo/kakao/data/local/SearchMetaData.kt
@@ -10,10 +10,10 @@ import java.io.Serializable
 data class SearchMetaData(
 
         @SerializedName("total_count")
-        var totalCount: Double,
+        var totalCount: Int,
 
         @SerializedName("pageable_count")
-        var pageableCount: Double,
+        var pageableCount: Int,
 
         @SerializedName("is_end")
         var isEnd: Boolean
